+++ library/stdio_vfscanf.c
@@ -1,5 +1,5 @@
 /*
- * $Id: stdio_vfscanf.c,v 1.15 2005-05-14 10:52:31 obarthel Exp $
+ * $Id: stdio_vfscanf.c,v 1.16 2005-05-29 08:19:36 obarthel Exp $
  *
  * :ts=4
  *
@@ -1107,14 +1107,14 @@ vfscanf(FILE *stream, const char *format, va_list arg)
 											double divisor;
 
 											/* A negative exponent means division. */
-											divisor = pow(radix,(double)exponent);
+											divisor = pow((double)radix,(double)exponent);
 											if(divisor != 0.0)
 												sum = sum / divisor;
 										}
 										else
 										{
 											/* A positive exponent means multiplication. */
-											new_sum = sum * pow(radix,(double)exponent);
+											new_sum = sum * pow((double)radix,(double)exponent);
 											if(new_sum >= sum)
 												sum = new_sum;
 											else
